requirements.txt
# Requirements for the additional code
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.9




# Requirements for scrapper.py
nltk==3.7
langchain-google-genai==0.0.14  # Ensure to replace with the correct version you are using
python-dotenv==1.0.0
google-generativeai==0.2.2  # Ensure to replace with the correct version you are using
